+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Bug fixes:
 
 - Return correct error code when killed
 
+New features:
+
+- Improve error logging by printing stack trace
+
 # 0.1.1 (2020-08-16)
 
 Bug fixes:
@@ -11,7 +15,7 @@ Bug fixes:
 - Fix bug where default values of config options are not used; the values are
   left undefined if they are not set
 
-New Features:
+New features:
 
 - Improve configurability of username and display names
 - Add script to change mattermost usernames of puppets